+++ app/src/main/java/com/example/interntasks_2a/case2/ListFragment.kt
@@ -12,6 +12,7 @@ import com.example.interntasks_2a.R
 import com.example.interntasks_2a.Weather
 import com.example.interntasks_2a.WeatherAdapter
 import com.example.interntasks_2a.databinding.FragmentListBinding
+import com.example.interntasks_2a.weatherList
 
 class ListFragment : Fragment() {
 
@@ -26,17 +27,6 @@ class ListFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
-        val weatherList = listOf(
-            Weather(10, 5, 15, "İstanbul", "Güneşli"),
-            Weather(15, 10, 20, "Ankara", "Bulutlu"),
-            Weather(20, 15, 25, "İzmir", "Yağmurlu"),
-            Weather(25, 20, 30, "Antalya", "Karlı"),
-            Weather(30, 25, 35, "Bursa", "Güneşli"),
-            Weather(35, 30, 40, "Konya", "Bulutlu"),
-            Weather(40, 35, 45, "Diyarbakır", "Yağmurlu"),
-            Weather(45, 40, 50, "Edirne", "Karlı")
-        )
 
         val adapter = WeatherAdapter(weatherList) { weather -> addBundle(weather) }
         binding.recyclerViewWeather.adapter = adapter
